Refactor methods on modules
