+++ src/js/maps.js
@@ -26,8 +26,28 @@ const geocodeCityLocation = () => {
   });
 };
 
+export const getCityLocation = async () => {
+  const city = document.querySelector('.city-name').value;
+  const GoogleServer = 'https://maps.googleapis.com/maps/api/geocode/json?address={city}&key={geo_key}';
+  const GeoUrl = GoogleServer
+    .replace('{geo_key}', process.env.GOOGLE_GEO_API_KEY)
+    .replace('{city}', city)
+    .replace(' ', '%20');
+  const response = await fetch(GeoUrl, { method: 'GET' });
+  const data = await response.json();
+  cityLocation.lat = data.results[0].geometry.location.lat;
+  cityLocation.lng = data.results[0].geometry.location.lng;
+  events.publish('cityLocated', cityLocation);
+};
+
+export const importGoogleSrc = () => {
+  const script = document.createElement('script');
+  script.src = `https://maps.googleapis.com/maps/api/js?key=${process.env.GOOGLE_NAM_API_KEY}&libraries=places`;
+  script.defer = true;
+  return script;
+};
+
 export const getDeviceLocation = () => {
-  console.warn('START GET LOCATION');
   navigator.geolocation.getCurrentPosition(position => {
     const pos = {
       lat: position.coords.latitude,
